fix(shape2 shape3) syncronized math (previously moved)
