--- src/core/math/vec3/canonicalize.js
@@ -1,8 +1,0 @@
-const fromValues = require('./fromValues');
-const { quantizeForSpace } = require('../utils')
-
-const canonicalize = (vector) => fromValues(quantizeForSpace(vector[0]),
-                                            quantizeForSpace(vector[1]),
-                                            quantizeForSpace(vector[2]))
-
-module.exports = canonicalize
