+++ src/core/shape2/clone.js
@@ -1,5 +1,5 @@
 const create = require('./create')
-const mat4 = require('../math/mat4')
+const mat4 = require('../../math/mat4')
 const geom2 = require('../geometry/geom2')
 
 /** create a copy/clone of the input shape
