+++ src/core/shape2/rotateY.js
@@ -1,7 +1,7 @@
 const transform = require('./transform')
 
-const fromYRotation = require('../math//mat4/fromXRotation')
-const {degToRad} = require('../math//utils')
+const fromYRotation = require('../../math/mat4/fromXRotation')
+const {degToRad} = require('../../math/utils')
 
 function rotateY (deg, shape2) {
   return transform(fromYRotation(degToRad(deg)), shape2)
