+++ src/core/shape3/translate.js
@@ -1,5 +1,5 @@
 const transform = require('./transform')
-const mat4 = require('../math/mat4')
+const mat4 = require('../../math/mat4')
 
 /** return a clone of the input shape, translated by the given Vec3
  * the clone's geometry still points to the original's, but the transforms are changed
